refactor(web): restructure ChatsPage component for improved modularity and readability
- Extracted various functionalities into dedicated components: ChatHeader, ChatList, ChatSearchBar, and ChatView.
- Simplified state management by utilizing the useChatPage hook for better organization of chat-related logic.
- Enhanced user experience with a cleaner layout and improved search functionality.
- Removed unused imports and redundant code to streamline the component.
- Ensured all changes adhere to TypeScript and linting standards.
