+++ apps/web/src/components/chats/ChatSearchBar.tsx
@@ -1,0 +1,37 @@
+'use client';
+
+import { cn } from '@babylon/shared';
+import { Search, X } from 'lucide-react';
+
+interface ChatSearchBarProps {
+  value: string;
+  onChange: (value: string) => void;
+}
+
+export function ChatSearchBar({ value, onChange }: ChatSearchBarProps) {
+  return (
+    <div className="relative mb-2 px-4">
+      <Search className="-translate-y-1/2 absolute top-1/2 left-7 h-4 w-4 text-muted-foreground" />
+      <input
+        type="text"
+        placeholder="Search conversations..."
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        className={cn(
+          'w-full rounded-lg py-2 pr-9 pl-9 text-sm',
+          'message-input bg-sidebar-accent/50',
+          'text-foreground placeholder:text-muted-foreground',
+          'outline-none'
+        )}
+      />
+      {value && (
+        <button
+          onClick={() => onChange('')}
+          className="-translate-y-1/2 absolute top-1/2 right-6 flex h-6 w-6 items-center justify-center rounded-md transition-colors hover:bg-muted-foreground/20"
+        >
+          <X className="h-4 w-4 text-foreground" />
+        </button>
+      )}
+    </div>
+  );
+}
