+++ apps/web/src/components/chats/index.ts
@@ -1,0 +1,12 @@
+export { ChatHeader } from './ChatHeader';
+export { ChatList } from './ChatList';
+export { ChatListItem } from './ChatListItem';
+export { ChatSearchBar } from './ChatSearchBar';
+export { ChatView } from './ChatView';
+export { ChatViewHeader } from './ChatViewHeader';
+export { FeedbackMessages } from './FeedbackMessages';
+export { MessageBubble } from './MessageBubble';
+export { MessageInput } from './MessageInput';
+export { MessageList } from './MessageList';
+export { useChatPage } from './hooks/useChatPage';
+export * from './types';
